refactor: Quick task modal
